fixed DataBase Model and more
fixed connection close
fixed edge class
added creating table script
added to user option to leave group

+++ src/Model/Edge.java
@@ -4,9 +4,9 @@ public class Edge {
 	private String station1;
 	private String station2;
 	private String city;
-	private int distance;
+	private float distance;
 	
-	public Edge(String station1, String station2, String city, int distance) {
+	public Edge(String station1, String station2, String city, float distance) {
 		super();
 		this.station1 = station1;
 		this.station2 = station2;
@@ -33,10 +33,10 @@ public class Edge {
 	public void setCity(String city) {
 		this.city = city;
 	}
-	public int getDistance() {
+	public float getDistance() {
 		return distance;
 	}
-	public void setDistance(int distance) {
+	public void setDistance(float distance) {
 		this.distance = distance;
 	}
